rename chain-id files
